Rename styles/NxTable/NxTablePage to styles/NxTable/NxTabeStylePage to avoid confusion
