+++ gallery/src/styles/NxTable/NxTableStylePage.tsx
